+++ src/data/duke.txt
@@ -1,3 +1,3 @@
-D | 0 | oop | 2019-12-03 | 5pm
-D | 1 | ff | 2019-03-21 | 12pm
-E | 1 | dfweA | 2018-02-21 | 5pm
+D | 0 |  return book | 2019-03-21 |  5pm
+E | 0 | returnhorse | 2017-02-01 |  3am
+T | 0 |  DWEDWAEW
