+++ src/main/java/TaskList.java
@@ -1,0 +1,18 @@
+import java.time.LocalDate;
+import java.util.ArrayList;
+
+public class TaskList {
+    public ArrayList<Task> tasks;
+
+    public TaskList() {
+        tasks = null;
+    }
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public ArrayList<Task> getTasks() {
+        return tasks;
+    }
+}
