+++ text-ui-test/input.txt
@@ -1,6 +1,6 @@
 todo borrow book
 list
-deadline return book /by Sunday
-event project meeting /at Mon 2-4pm
+deadline return book /by 2019-03-21 time: 5pm
+event return horse /at 2017-02-01 time: 3am
 list
 bye
